# Improve performance of instantiating exceptions/errors