@nielsdos 2025-04-27
> I haven't measured recently but I know when people say that exceptions are expensive, they also generally factor in walking the call stack.

Sure, but I'm also convinced that can be improved performance-wise :slightly_smiling_face: 

> (...) but it has a meaningful benchmark in it that you could run that's roughly 85 call frames deep, which matched a real performance issue one of our customers had.

I might take a look at using that as a benchmark to improve backtrace construction, if that's something I would work on in the future.

> And maybe adjust the "up to 50% faster" wording to be a bit more realistic :)

Sure