@arnaud-lb 2025-04-29
This is great. The same technique could possibly be used for other internal classes that have standard properties.

Just some thoughts, not meant to block this PR:

We could use stub files to generate offset macros/constants, and expose `zend_obj_prop_num_checked()`.

Since Exceptions are almost user land classes (they have no custom handlers except clone_obj), I'm wondering how it would perform if we actually declared these classes in user land (c.f. https://github.com/php/php-src/pull/18204).